Initial node-ification and AMD-ification.
Wrap modules in AMD wrappers for browser-safety; move the files around
a bit to conform to standard node/volo package organization.  Add 'lzmajs'
binary (not yet working) and make test suite mocha-compatible.  Add README.
