+++ README.md
@@ -1,0 +1,19 @@
+# lzmajs
+
+`lzmajs` is a fast pure-JavaScript implementation of LZMA
+compression/decompression.  It was originally written by Gary Linscott;
+C. Scott Ananian cleaned up the source code and packaged it for `node`
+and `volo`.
+
+## How to install
+
+```
+npm install lzmajs
+```
+or
+```
+volo add cscott/lzmajs
+```
+
+## Usage
+## Documentation
