--- build
@@ -1,2 +1,0 @@
-jslint encoder.js
-node test.js
