+++ docs/optimizing javascript
