+++ lib/lzma.js
@@ -1,3 +1,6 @@
+if (typeof define !== 'function') { var define = require('amdefine')(module); }
+define([],function(){
+'use strict';
 
 var LZMA = LZMA || {};
 
@@ -91,7 +94,7 @@ LZMA.RangeDecoder.prototype.init = function(){
 
   this._code = 0;
   this._range = -1;
-  
+
   while(i --){
     this._code = (this._code << 8) | this._stream.readByte();
   }
@@ -509,4 +512,7 @@ LZMA.decompress = function(properties, inStream, outStream, outSize){
   return true;
 };
 
-exports.LZMA = LZMA;
+    return {
+        LZMA: LZMA
+    };
+});
