+++ package.json
@@ -1,0 +1,30 @@
+{
+  "name": "lzmajs",
+  "version": "0.9.0",
+  "author": "Gary Linscott, C. Scott Ananian",
+  "description": "a pure-JavaScript implementation of LZMA de/compression.",
+  "keywords": ["lzma", "compression", "decompression"],
+  "main": "main",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/cscott/lzmajs.git"
+  },
+  "bin": {
+    "lzmajs": "./bin/lzmajs"
+  },
+  "directories": {
+    "test": "test"
+  },
+  "dependencies": {
+    "amdefine": "~0.0.4",
+    "commander": "~1.1.1"
+  },
+  "devDependencies": {
+    "mocha": "~1.8.1"
+  },
+  "scripts": {
+    "test": "mocha"
+  },
+  "amd": {},
+  "volo": {}
+}
